style(java): 优化代码格式和注释规范
- 统一 AbstractAudioModelConfig 中的字段注释格式
- 调整 AbstractBaseConsumer 中 import 语句顺序
- 规范 AbstractCharacterChunkingStrategy 注释格式
- 修复 AbstractKeepEdgeMaskingStrategy 中参数注释间距
- 格式化 AbstractMqttClientApi 类声明换行
- 标准化 AbstractObjectStorageEngine 字段注释格式
- 优化 AbstractObjectStorageEngineTest 中方法参数对齐
- 调整 AbstractOcrProviderTest 中 import 语句顺序
- 重新排列 AbstractStreamConsumer 中 import 语句位置
- 规范 AbstractStreamConsumer 内部类字段注释格式
- 标准化 AbstractStreamConsumer 构造器参数注释格式
- 统一 AbstractStreamConsumer 成员变量注释格式
- 修复 AbstractStreamConsumer 中代码缩进问题
- 格式化 AbstractTenancyStrategy 中 import 语句顺序
- 优化 AbstractVectorService 中 import 语句组织
- 调整 AbstractVectorService 成员变量顺序和注释格式
- 规范 AbstractVectorService 方法参数对齐
- 标准化 AccessLogAspect 中 import 语句顺序
- 修复 AccessLogAspect 方法参数注释格式
- 优化 AccessLogService 接口注释格式
- 标准化 AccessLogServiceImpl 方法参数注释格式
- 清理 AccessScope 类中多余注释空行
- 调整 additional-spring-configuration-metadata.json 格式

+++ atlas-richie-component/atlas-richie-component-cache/src/main/java/cn/richie696/component/cache/local/config/LocalCacheAutoConfiguration.java
@@ -47,7 +47,9 @@ import static cn.richie696.component.cache.enums.L2CachingRegion.ACCESS_LOG;
 @EnableConfigurationProperties({LocalCacheProperties.class})
 public class LocalCacheAutoConfiguration {
 
-    /** 默认构造函数，供 Spring 实例化使用。 */
+    /**
+     * 默认构造函数，供 Spring 实例化使用。
+     */
     public LocalCacheAutoConfiguration() {
     }
 
@@ -98,7 +100,7 @@ public class LocalCacheAutoConfiguration {
                 .setStoreByValue(cacheDefinition.isStoreByValue())
                 .setStatisticsEnabled(cacheDefinition.isStatisticsEnabled());
 
-        Duration expiryDuration = new Duration(cacheDefinition.getExpiryUnit(), cacheDefinition.getExpiry());
+        Duration expiryDuration = new Duration(cacheDefinition.getExpiryUnit(), cacheDefinition.getExpiry().longValue());
         switch (cacheDefinition.getExpiryPolicy()) {
             case CREATED -> configuration.setExpiryPolicyFactory(CreatedExpiryPolicy.factoryOf(expiryDuration));
             case ACCESSED -> configuration.setExpiryPolicyFactory(AccessedExpiryPolicy.factoryOf(expiryDuration));
